requirements.txt
# ProjectA WP Multi-Updater - Python Dependencies

# YAML parsing for inventory files
PyYAML>=6.0.1

# HTTP requests for Cloudways API integration (optional)
requests>=2.31.0

# GUI Dependencies
Flask>=3.0.0
Flask-SocketIO>=5.3.0
python-socketio>=5.10.0

# No other external dependencies required for CLI!
# The CLI tool uses Python 3.10+ stdlib for:
# - subprocess (ssh/scp)
# - sqlite3 (results database)
# - csv (job files and reports)
# - concurrent.futures (ThreadPoolExecutor)
# - argparse (CLI)
# - logging (structured output)
